backend_py/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
cryptography>=41.0.5
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# HTTP and API
httpx>=0.25.1
python-dotenv>=1.0.0
starlette>=0.31.1
python-dateutil>=2.8.2

# AI and NLP
perplexity==0.0.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
asyncio>=3.4.3
aiohttp>=3.8.6

# Production
gunicorn>=21.2.0
python-json-logger>=2.0.7
prometheus-client>=0.18.0
sentry-sdk>=1.34.0
psutil>=5.9.6
docker>=7.0.0
bcrypt>=4.1.1

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0